Move `VaryValue`, `HeaderFieldCombiner`, and `HeaderValuesEqual` to `internal/xhttp`
